refactor: split autoApply helpers
